+++ lib/src/code_generators/swagger_additions_generator.dart
@@ -39,9 +39,9 @@ class SwaggerAdditionsGenerator {
       final className =
           "${getClassNameFromFileName(key.split('/').last).camelCase}$converterClassEnding";
 
-      final fileName = key.split('/').last.replaceAll('.json', '.swagger');
+      final fileName = getFileNameBase(key);
       maps.writeln('  ...$className,');
-      imports.add("import '${fileName.replaceAll('-', '_')}.dart';");
+      imports.add("import '$fileName.swagger.dart';");
     });
 
     final mapping = '''