+++ lib/src/swagger_models/swagger_components.dart
@@ -19,10 +19,10 @@ class SwaggerComponents {
   @JsonKey(name: 'schemas', defaultValue: {})
   Map<String, SwaggerSchema> schemas;
 
-  @JsonKey(name: 'responses', defaultValue: {}, fromJson: mapResponses)
+  @JsonKey(name: 'responses', fromJson: _mapResponses)
   Map<String, SwaggerSchema> responses;
 
-  @JsonKey(name: 'requestBodies', defaultValue: {}, fromJson: mapResponses)
+  @JsonKey(name: 'requestBodies', fromJson: _mapResponses)
   Map<String, SwaggerSchema> requestBodies;
 
   Map<String, dynamic> toJson() => _$SwaggerComponentsToJson(this);
@@ -31,7 +31,7 @@ class SwaggerComponents {
       _$SwaggerComponentsFromJson(json);
 }
 
-Map<String, SwaggerSchema> mapResponses(Map<String, dynamic>? json) {
+Map<String, SwaggerSchema> _mapResponses(Map<String, dynamic>? json) {
   var results = <String, SwaggerSchema>{};
 
   if (json == null) {
@@ -42,18 +42,17 @@ Map<String, SwaggerSchema> mapResponses(Map<String, dynamic>? json) {
     final content =
         (value as Map<String, dynamic>)['content'] as Map<String, dynamic>?;
 
-    final appJsonKey = content?.keys.firstWhere(
-            (element) => element.startsWith('application/json'),
-            orElse: () => '') ??
-        '';
+    Map<String, dynamic>? appJson;
 
-    if (appJsonKey.isNotEmpty) {
-      final appJson = content?[appJsonKey] as Map<String, dynamic>?;
+    if (content?.length == 1) {
+      appJson = content?.values.first as Map<String, dynamic>?;
+    } else {
+      appJson = content?['application/json'] as Map<String, dynamic>?;
+    }
 
-      if (appJson != null && appJson['schema'] != null) {
-        results[key] =
-            SwaggerSchema.fromJson(appJson['schema'] as Map<String, dynamic>);
-      }
+    if (appJson != null && appJson['schema'] != null) {
+      results[key] =
+          SwaggerSchema.fromJson(appJson['schema'] as Map<String, dynamic>);
     }
   });