+++ lib/src/swagger_models/swagger_root.dart
@@ -1,3 +1,4 @@
+import 'package:swagger_dart_code_generator/src/code_generators/constants.dart';
 import 'package:swagger_dart_code_generator/src/swagger_models/requests/swagger_request.dart';
 import 'package:swagger_dart_code_generator/src/swagger_models/requests/swagger_request_parameter.dart';
 import 'package:swagger_dart_code_generator/src/swagger_models/responses/swagger_schema.dart';
@@ -38,7 +39,7 @@ class SwaggerRoot {
   @JsonKey(name: 'schemes', defaultValue: [])
   List<String> schemes;
 
-  @JsonKey(name: 'paths', defaultValue: {}, fromJson: _mapPaths)
+  @JsonKey(name: 'paths', fromJson: _mapPaths)
   Map<String, SwaggerPath> paths;
 
   @JsonKey(name: 'definitions', defaultValue: {})
@@ -56,11 +57,16 @@ class SwaggerRoot {
       _$SwaggerRootFromJson(json);
 }
 
-Map<String, SwaggerPath> _mapPaths(Map<String, dynamic> paths) {
+Map<String, SwaggerPath> _mapPaths(Map<String, dynamic>? paths) {
+  if (paths == null) {
+    return {};
+  }
+
   return paths.map((path, pathValue) {
     final value = pathValue as Map<String, dynamic>;
     final parameters = value['parameters'] as List<dynamic>?;
-    value.removeWhere((key, value) => key == 'parameters');
+    value.removeWhere(
+        (key, value) => !supportedRequestTypes.contains(key.toLowerCase()));
 
     return MapEntry(
       path,