+++ lib/swagger_dart_code_generator.dart
@@ -25,8 +25,7 @@ Map<String, List<String>> _generateExtensions(GeneratorOptions options) {
   final result = <String, List<String>>{};
 
   filesList.forEach((FileSystemEntity element) {
-    final name =
-        element.path.split('/').last.split('.').first.replaceAll('-', '_');
+    final name = getFileNameBase(element.path);
     result[element.path] = <String>[
       '${options.outputFolder}$name$_outputFileExtension',
       '${options.outputFolder}$name$_outputEnumsFileExtension',
@@ -61,7 +60,7 @@ class SwaggerDartCodeGenerator implements Builder {
   Future<void> build(BuildStep buildStep) async {
     final fileNameWithExtension =
         buildStep.inputId.pathSegments.last.replaceAll('-', '_');
-    final fileNameWithoutExtension = fileNameWithExtension.split('.').first;
+    final fileNameWithoutExtension = getFileNameBase(fileNameWithExtension);
 
     final contents = await buildStep.readAsString(buildStep.inputId);
 
@@ -139,7 +138,7 @@ $imports
 
 ${options.buildOnlyModels ? '' : requests}
 
-${options.withConverter ? converter : ''}
+${options.withConverter && !options.buildOnlyModels ? converter : ''}
 
 $models
 
@@ -147,7 +146,7 @@ $responses
 
 $requestBodies
 
-$customDecoder
+${options.buildOnlyModels ? '' : customDecoder}
 
 $dateToJson
 """;
@@ -178,7 +177,7 @@ $dateToJson
 
     await buildStep.writeAsString(indexAssetId, _formatter.format(imports));
 
-    if (options.withConverter) {
+    if (options.withConverter && !options.buildOnlyModels) {
       final mappingAssetId =
           AssetId(inputId.package, '${options.outputFolder}$_mappingFileName');